+++ DependencyInjection/c975LPaymentExtension.php
@@ -29,22 +29,5 @@ class c975LPaymentExtension extends Extension
             new FileLocator(__DIR__.'/../Resources/config')
         );
         $loader->load('services.yml');
-
-        $configuration = new Configuration();
-        $processedConfig = $this->processConfiguration($configuration, $configs);
-
-        $container->setParameter('c975_l_payment.live', $processedConfig['live']);
-        $container->setParameter('c975_l_payment.site', $processedConfig['site']);
-        $container->setParameter('c975_l_payment.gdpr', $processedConfig['gdpr']);
-        $container->setParameter('c975_l_payment.defaultCurrency', strtoupper($processedConfig['defaultCurrency']));
-        $container->setParameter('c975_l_payment.vat', $processedConfig['vat'] * 100);
-        $container->setParameter('c975_l_payment.stripeFeePercentage', $processedConfig['stripeFeePercentage']);
-        $container->setParameter('c975_l_payment.stripeFeeFixed', $processedConfig['stripeFeeFixed']);
-        $container->setParameter('c975_l_payment.timezone', $processedConfig['timezone']);
-        $container->setParameter('c975_l_payment.database', $processedConfig['database']);
-        $container->setParameter('c975_l_payment.image', $processedConfig['image']);
-        $container->setParameter('c975_l_payment.zipCode', $processedConfig['zipCode']);
-        $container->setParameter('c975_l_payment.alipay', $processedConfig['alipay']);
-        $container->setParameter('c975_l_payment.roleNeeded', $processedConfig['roleNeeded']);
     }
 }