+++ ChangeLog.md
@@ -1,5 +1,22 @@
 # Changelog
 
+v2.0
+----
+**Upgrading from v1.x? Check UPGRADE.md**
+- Created branch 1.x (31/08/2018)
+- Updated composer.json (01/09/2018)
+- Updated `README.md` (01/09/2018)
+- Added `UPGRADE.md` (01/09/2018)
+- Added `bundle.yaml` (01/09/2018)
+- Removed declaration of parameters in Configuration class as they are end-user parameters and defined in c975L/ConfigBundle (01/09/2018)
+- Added Route `payment_config` (01/09/2018)
+- Added shortcut `$paymentService->getParameter()` (01/09/2018)
+- Removed calls of `$container->getParameter()` (01/09/2018)
+
+
+v1.x
+====
+
 v1.16.2
 -------
 - Fixed Voter constants (31/08/2018)